# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/customers": {
            "get": {
                "description": "Returns full list of existing customers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "list-customers"
                ],
                "summary": "Get list of customers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "first name search",
                        "name": "firstName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "last name search",
                        "name": "lastName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "list length",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "list offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Customer"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create customer object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create-customer"
                ],
                "summary": "Create customer",
                "parameters": [
                    {
                        "description": "Customer details",
                        "name": "customerDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Customer"
                        }
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/customers/{customerID}": {
            "get": {
                "description": "Returns customer details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get-customer"
                ],
                "summary": "Get customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Customer"
                        }
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit customer details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "edit-customer"
                ],
                "summary": "Edit customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New customer details",
                        "name": "customerDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.EditCustomerDetailsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Customer"
                        }
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete customer details and it's relations by ID",
                "tags": [
                    "delete-customer"
                ],
                "summary": "Delete customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/customers/{customerID}/purchases": {
            "get": {
                "description": "Returns full list of purchases for a specific customer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get-customer-purchases"
                ],
                "summary": "Get list of purchases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Purchase"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new purchase for a customer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create-customer-purchase"
                ],
                "summary": "Create a purchase for a customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Purchase details",
                        "name": "purchaseDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreatePurchaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Purchase"
                        }
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/customers/{customerID}/purchases/{purchaseID}": {
            "put": {
                "description": "Updates a purchase for a customer by ID",
                "tags": [
                    "update-customer-purchase"
                ],
                "summary": "Update a purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Purchase ID",
                        "name": "purchaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New purchase details",
                        "name": "customerDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.EditPurchaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Purchase"
                        }
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a purchase by ID",
                "tags": [
                    "delete-customer-purchase"
                ],
                "summary": "Delete a purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Purchase ID",
                        "name": "purchaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/customers/{customerID}/repairs/{repairID}": {
            "delete": {
                "description": "Deletes a repair by ID",
                "tags": [
                    "delete-customer-repair"
                ],
                "summary": "Delete a repair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repair ID",
                        "name": "repairID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "IMPLEMENTED BUT DOCS TODO",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.Customer": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "telephone_number"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "telephone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "server.CreateCustomerRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "telephone_number"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "telephone_number": {
                    "type": "string"
                }
            }
        },
...
        "server.EditCustomerDetailsRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "telephone_number"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "telephone_number": {
                    "type": "string"
                }
            }
        },
...
    }
}